Reformat code added the possibility to remove a selected recipe Added the generate button and message

+++ src/model/ingredient.rs
@@ -2,5 +2,5 @@
 #[derive(Debug)]
 pub struct Ingredient {
     pub name: String,
-    pub quantity: u8,
+    pub quantity: f32,
 }
